pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matrix-synapse-automator"
version = "0.22"
description = "Automates the repetitive processes inside the Matrix Synapse server"
readme = "README.md"
authors = [{name = "Insight Builder", email = "insighhacker21@gmail.com"}]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
]
keywords = ["matrix synapse", "homeserver", "matrix server"]
dependencies = [
    "matrix-nio[e2e]",
    "markdown",
    "emoji",
    "python_magic",
    "dbus-python",
    "notify2",
    "urllib",
    "pyxdg"
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/insightbuilder/matrix_synapse_automator_scripts"

[project.scripts]
matrix_automator = "matrix_synapse_automator_scripts.__main__:main"

[tools.setuptools]
py-modules = ["api_viewer", "automation_helper", "gbot_msgsender_script", "invite_acceptor_script", "invite_sender_script", "listen_and_ai_reply", "messages_sync_script", "message_sender", "room_leave_forget"]